pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "idpkit"
version = "2.0.0"
description = "IDP Kit — Intelligent Document Processing Toolkit & AI Agent"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "IDP Kit Team"},
]
keywords = ["document-processing", "rag", "ai-agent", "pdf", "nlp", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: General",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # Engine (existing PageIndex deps)
    "openai>=1.101.0",
    "pymupdf>=1.26.4",
    "PyPDF2>=3.0.1",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.11.0",
    "pyyaml>=6.0.2",
    # LLM abstraction (multi-provider)
    "litellm>=1.40.0",
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "itsdangerous>=2.1.0",
    # Templates & static files
    "jinja2>=3.1.0",
    "aiofiles>=24.0.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    # Document parsing
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "openpyxl>=3.1.0",
    "python-pptx>=1.0.0",
    # Utilities
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
postgres = ["asyncpg>=0.29.0"]
ocr = ["pytesseract>=0.3.10"]
graph = ["networkx>=3.0"]
all = [
    "asyncpg>=0.29.0",
    "pytesseract>=0.3.10",
    "networkx>=3.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
idpkit = "idpkit.cli:main"

[project.urls]
Homepage = "https://github.com/idpkit/idpkit"
Documentation = "https://idpkit.dev"
Repository = "https://github.com/idpkit/idpkit"

[tool.setuptools.packages.find]
include = ["idpkit*", "pageindex*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
